requirements.txt
Flask==2.0.2
google-cloud-bigquery==2.30.1
google-auth==1.35.0
gunicorn==20.1.0
